+++ app/src/main/java/com/example/recipeapp/model/Category.kt
@@ -1,0 +1,8 @@
+package com.example.recipeapp.model
+
+data class Category(
+    val id: Int,
+    val title: String,
+    val description: String,
+    val imageUrl: String,
+)
