+++ app/src/main/java/com/example/recipeapp/model/Ingredient.kt
@@ -1,0 +1,7 @@
+package com.example.recipeapp.model
+
+data class Ingredient(
+    val quantity: String,
+    val unitOfMeasure: String,
+    val description: String,
+)
