+++ app/src/main/java/com/example/recipeapp/model/Recipe.kt
@@ -1,0 +1,9 @@
+package com.example.recipeapp.model
+
+data class Recipe(
+    val id: Int,
+    val title: String,
+    val ingredients: List<Ingredient>,
+    val method: String,
+    val imageUrl: String,
+)
