+++ app/src/main/res/layout/activity_main.xml
@@ -38,6 +38,7 @@
             android:layout_marginEnd="@dimen/small_20"
             android:layout_weight="1"
             android:backgroundTint="@color/btn_categories_color"
+            android:contentDescription="@string/categoriesBtn"
             android:text="@string/title_categories" />
 
         <Button
@@ -48,6 +49,7 @@
             android:layout_marginStart="@dimen/small_20"
             android:layout_weight="1"
             android:backgroundTint="@color/btn_favorites_color"
+            android:contentDescription="@string/favoritesBtn"
             android:text="@string/title_favorites"
             app:icon="@drawable/ic_heart_empty"
             app:iconGravity="end" />
